Extract NFT plugin in its own file
