+++ src/drivers/storage/BundlrStorageDriver.ts
@@ -25,7 +25,9 @@ export interface BundlrOptions {
 export const bundlrStorage = (options: BundlrOptions = {}): MetaplexPlugin => ({
   install(metaplex: Metaplex) {
     metaplex.setStorage(new BundlrStorageDriver(metaplex, options));
-    metaplex.register(planUploadMetadataOperation, planUploadMetadataUsingBundlrOperationHandler);
+    metaplex
+      .operations()
+      .register(planUploadMetadataOperation, planUploadMetadataUsingBundlrOperationHandler);
   },
 });
